VideoPress: Move the caption track REST handlers into their endpoint
The handlers, their private helpers (permission resolution, content
sanitization, response shaping), the exposed meta-key list, and the
request size caps had no consumer outside the wpcom/v2 controller, so
they live there now. Caption_Tracks keeps the storage concern: the
CPT, its meta registration, and the value sanitizers the meta layer
and the endpoint share.
